Refactor backend structure
