+++ backend/documents/__init__.py
@@ -1,0 +1,1 @@
+# Subpackage for document models and factories
